feat: [console-lite] - top nav-bar - fixes and improvements
