+++ apps/simple-trading-app/src/app/app.tsx
@@ -8,16 +8,11 @@ import {
   VegaManageDialog,
   VegaWalletProvider,
 } from '@vegaprotocol/wallet';
-import { VegaWalletConnectButton } from './components/vega-wallet-connect-button';
-import { ThemeSwitcher } from '@vegaprotocol/ui-toolkit';
 import { Connectors } from './lib/vega-connectors';
 import '../styles.scss';
 import { AppLoader } from './components/app-loader';
-import Logo from './components/icons/logo';
-import Video from './components/video';
-import Comet from './components/icons/comet';
+import Header from './components/header';
 import { Main } from './components/main';
-import { DrawerToggle, DRAWER_TOGGLE_VARIANTS } from './components/drawer';
 import { useLocation } from 'react-router-dom';
 
 function App() {
@@ -43,40 +38,11 @@ function App() {
           <VegaWalletProvider>
             <AppLoader>
               <div className="max-h-full min-h-full dark:bg-black dark:text-white-60 bg-white text-black-60 grid grid-rows-[min-content,1fr]">
-                <div className="flex items-stretch p-16 bg-black dark:bg-white text-white-60">
-                  <div className="absolute top-0 right-[200px] w-[500px] h-[100px] z-0 hidden md:block">
-                    <Video />
-                    <div
-                      id="swarm"
-                      className="absolute w-[500px] h-[100px] bg-black dark:bg-white"
-                    />
-                  </div>
-                  <Logo theme={theme} />
-                  <Comet />
-                  <div className="flex items-center gap-4 ml-auto mr-8">
-                    <VegaWalletConnectButton
-                      setConnectDialog={(open) =>
-                        setVegaWallet((x) => ({ ...x, connect: open }))
-                      }
-                      setManageDialog={(open) =>
-                        setVegaWallet((x) => ({ ...x, manage: open }))
-                      }
-                    />
-                    <ThemeSwitcher
-                      onToggle={toggleTheme}
-                      className="-my-4"
-                      sunClassName="text-white"
-                    />
-                  </div>
-                  <DrawerToggle
-                    onToggle={onToggle}
-                    variant={DRAWER_TOGGLE_VARIANTS.OPEN}
-                    className="xs:py-32 xs:px-16 xs:text-white xs:hover:text-blue"
-                  />
-                </div>
-
+                <Header
+                  setVegaWallet={setVegaWallet}
+                  toggleTheme={toggleTheme}
+                />
                 <Main isMenuOpen={menuOpen} onToggle={onToggle} />
-
                 <VegaConnectDialog
                   connectors={Connectors}
                   dialogOpen={vegaWallet.connect}
