+++ apps/simple-trading-app/src/app/components/header/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './header';
